Add files via upload
The database application is for an auto shop business. The application allows users to place an appointment for auto service and it also allows the shop administrators to view the placed orders, store the employees' and customers' information. Customers can also make payments. The queries in the Java program are used to manipulate this database to extract useful information like placed services, payments information and, the total spent amount by each customer.